requirements.txt
python-telegram-bot==20.5
requests==2.31.0
python-telegram-bot[job-queue]